requirements.txt
aio-pika==8.1.1
fastapi==0.83.0
gunicorn==20.0.4
httptools==0.5.0
httpx==0.19.0
jsonschema==3.2.0
neo4j==5.0.1
pamqp==3.2.0
pydantic==1.10.2
pytest==7.1.2
pytest-asyncio==0.15.1
pytest-dotenv==0.5.2
pyyaml==6.0
reasoner-pydantic==4.0.5
redis~=3.5.3
requests==2.28.1
uvicorn==0.17.6
uvloop==0.17.0
opentelemetry-sdk==1.16.0
opentelemetry-instrumentation-fastapi==0.37b0
opentelemetry-exporter-jaeger==1.16.0
opentelemetry-instrumentation-httpx==0.37b0